moved PSF routines to psf_utils.py
